Mainly just some method definitions, file creations, and tweaking some post methods
